+++ src/panelWebView/components/BaseView.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { FolderInfo, PanelSettings } from '../../models';
+import { CustomScript, FolderInfo, PanelSettings } from '../../models';
 import { CommandToCode } from '../CommandToCode';
 import { MessageHelper } from '../../helpers/MessageHelper';
 import { Collapsible } from './Collapsible';
@@ -31,6 +31,12 @@ const BaseView: React.FunctionComponent<IBaseViewProps> = ({settings, folderAndF
     MessageHelper.sendMessage(CommandToCode.openPreview);
   };
 
+  const runBulkScript = (script: CustomScript) => {
+    MessageHelper.sendMessage(CommandToCode.runCustomScript, { title: script.title, script });
+  };
+
+  const customActions: any[] = (settings?.scripts || []).filter(s => s.bulk && (s.type === "content" || !s.type));
+
   return (
     <div className="frontmatter">
       <div className={`ext_actions`}>
@@ -42,6 +48,11 @@ const BaseView: React.FunctionComponent<IBaseViewProps> = ({settings, folderAndF
             <button onClick={initProject} disabled={settings?.isInitialized}>Initialize project</button>
             <button onClick={createContent} disabled={!settings?.isInitialized}>Create new content</button>
             <button onClick={openPreview} disabled={!settings?.preview?.host}>Open site preview</button>
+            {
+              customActions.map((script) => (
+                <button key={script.title} onClick={() => runBulkScript(script)}>{ script.title }</button>
+              ))
+            }
           </div>
         </Collapsible>